Hop la

+++ moteur/classes/Tacticien.java
@@ -65,7 +65,7 @@ public abstract class Tacticien extends Pion {
 	 * @return 
 	 */
 	@Override
-	protected float janken(Pion p) {
+	protected double janken(Pion p) {
 		return 0;
 	}
